move user response schema to user_schema
